+++ lib/src/features/auth/code/code_vm.dart
@@ -1,36 +1,44 @@
+import 'dart:async';
+
 import 'package:flutter/widgets.dart';
 import 'package:mobx/mobx.dart';
 import 'package:routemaster/routemaster.dart';
 import 'package:where_to_go_today/src/core/services/exceptions/server/server_error_exception.dart';
 import 'package:where_to_go_today/src/core/ui/base/view_model.dart';
 import 'package:where_to_go_today/src/core/ui/errors_handling/error_handler.dart';
-import 'package:where_to_go_today/src/features/auth/code/code_route.dart';
+import 'package:where_to_go_today/src/features/auth/code/code_vm_state.dart';
 import 'package:where_to_go_today/src/features/auth/register/register_route.dart';
 import 'package:where_to_go_today/src/features/auth/services/auth_bloc.dart';
 import 'package:where_to_go_today/src/features/auth/services/bloc/events/auth_event.dart';
 import 'package:where_to_go_today/src/features/auth/services/bloc/states/auth_state.dart';
-import 'package:where_to_go_today/src/features/auth/sign_in/sign_in_vm_state.dart';
 
-part 'sign_in_screen_vm.g.dart';
+part 'code_vm.g.dart';
 
-class SignInScreenVm = _SignInScreenVm with _$SignInScreenVm;
+class CodeVm = _CodeVm with _$CodeVm;
 
-abstract class _SignInScreenVm extends ViewModel with Store {
+abstract class _CodeVm extends ViewModel with Store {
+  static const int _resendingTimeout = 30;
   final BuildContext _context;
   final AuthBloc _bloc;
 
   @observable
-  SignInVmState vmState = SignInVmState.idle;
+  CodeVmState vmState = CodeVmState.idle;
 
   @observable
-  bool isPhoneValid = false;
+  bool isValidCode = false;
+
+  @observable
+  int countdown = 0;
 
   @computed
-  bool get isLoading => vmState == SignInVmState.loading;
+  bool get canRequestCode => countdown == 0;
 
-  String? _phone;
+  @computed
+  bool get isLoading => vmState == CodeVmState.loading;
 
-  _SignInScreenVm(
+  Timer? _timer;
+
+  _CodeVm(
     this._context,
     this._bloc, {
     required ErrorHandler errorHandler,
@@ -38,49 +46,55 @@ abstract class _SignInScreenVm extends ViewModel with Store {
     observeBloc<AuthState, AuthBloc>(_bloc, _handleStates);
   }
 
+  @override
+  void dispose() {
+    _timer?.cancel();
+    super.dispose();
+  }
+
   @action
-  void verifyPhone(String phone) {
-    isPhoneValid = phone.length == 12;
+  void startTimer() {
+    _startTimer();
+  }
+
+  @action
+  void validateCode(String? value) {
+    isValidCode = value?.length == 6;
   }
 
   @action
   Future<void> requestCode(String phone) async {
-    _phone = phone;
     _bloc.add(AuthEvent.sendPhone(phone));
+    _startTimer();
   }
 
   @action
-  void signInWithGoogle() {
-    _bloc.add(const AuthEvent.loginViaGoogle());
+  void sendCode(String code) {
+    _bloc.add(AuthEvent.sendOtp(code));
   }
 
-  @action
-  void signInWithFacebook() {
-    _bloc.add(const AuthEvent.loginViaFacebook());
-  }
-
-  @action
-  void signInWithVkontakte() {
-    _bloc.add(const AuthEvent.loginViaVkontakte());
+  void _startTimer() {
+    countdown = _resendingTimeout;
+    _timer = Timer.periodic(const Duration(seconds: 1), (timer) {
+      countdown--;
+      if (countdown == 0) timer.cancel();
+    });
   }
 
   void _handleStates(AuthState state) {
     if (state is AuthStateIdle) {
-      vmState = SignInVmState.loading;
+      vmState = CodeVmState.loading;
     } else if (state is AuthStateNeedOtp) {
-      vmState = SignInVmState.needOtp;
-      if (_phone != null) {
-        Routemaster.of(_context).replace('${CodeRoute.routeName}/$_phone');
-      }
-    } else if (state is AuthStateSuccessViaSocial) {
-      vmState = SignInVmState.successSocial;
+      vmState = CodeVmState.needOtp;
+    } else if (state is AuthStateSuccessViaOtp) {
+      vmState = CodeVmState.successOtp;
       Routemaster.of(_context).replace(RegisterRoute.routeName);
     } else if (state is AuthStateError) {
-      vmState = SignInVmState.error;
+      vmState = CodeVmState.error;
       _bloc.onError(AuthorizationException(), state.stackTrace);
-      vmState = SignInVmState.idle;
+      vmState = CodeVmState.idle;
     } else {
-      vmState = SignInVmState.idle;
+      vmState = CodeVmState.idle;
     }
   }
 }