+++ lib/src/features/auth/services/auth_bloc.dart
@@ -1,8 +1,16 @@
+import 'dart:async';
+
 import 'package:bloc/bloc.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter/services.dart';
 import 'package:where_to_go_today/src/core/services/base/can_throw_exception_bloc_mixin.dart';
+import 'package:where_to_go_today/src/core/services/exceptions/server/server_error_exception.dart';
 import 'package:where_to_go_today/src/features/auth/services/bloc/events/auth_event.dart';
 import 'package:where_to_go_today/src/features/auth/services/bloc/states/auth_state.dart';
+import 'package:where_to_go_today/src/features/auth/services/google/google_auth.dart';
+import 'package:where_to_go_today/src/features/auth/services/vk/vk_auth.dart';
+import 'package:where_to_go_today/src/features/authapi/models/requests/google_login_request.dart';
+import 'package:where_to_go_today/src/features/authapi/models/requests/vk_login_request.dart';
 import 'package:where_to_go_today/src/features/authservices/repository/auth_repository.dart';
 
 /// Сервис позволяющий:
@@ -15,67 +23,121 @@ import 'package:where_to_go_today/src/features/authservices/repository/auth_repo
 class AuthBloc extends Bloc<AuthEvent, AuthState>
     with CanThrowExceptionBlocMixin {
   final AuthRepository authRepository;
+  final GoogleAuth googleAuth;
+  final VKAuth vkAuth;
 
-  bool firstSending = true;
+  AuthBloc({
+    required this.authRepository,
+    required this.googleAuth,
+    required this.vkAuth,
+  }) : super(const AuthState.init()) {
+    on<AuthEventSendPhone>(_onSendPhone);
+    on<AuthEventSendOtp>(_onSendOtp);
+    on<AuthEventLoginViaFacebook>(_onLoginViaFacebook);
+    on<AuthEventLoginViaVkontakte>(_onLoginViaVkontakte);
+    on<AuthEventLoginViaGoogle>(_onLoginViaGoogle);
+    on<AuthEventRegister>(_onRegister);
+    on<AuthEventLogout>(_onAuthEventLogout);
+  }
 
-  AuthBloc({required this.authRepository}) : super(const AuthState.init()) {
-    on<AuthEventSendPhone>((event, emit) async {
-      if (firstSending) {
-        emit(const AuthState.idle());
-        // TODO(any): handle first incoming `AuthEventSendPhone` event
-        emit(const AuthState.needOtp());
-        firstSending = false;
-      } else {
-        emit(const AuthState.idle());
-        // TODO(any): handle next incoming `AuthEventSendPhone` event
-        emit(const AuthState.success());
-      }
-    });
+  FutureOr<void> _onSendPhone(
+    AuthEventSendPhone _,
+    Emitter<AuthState> emit,
+  ) async {
+    emit(const AuthState.idle());
+    try {
+      // TODO(any): handle incoming `AuthEventSendPhone` event
+      await Future.delayed(const Duration(seconds: 1), () {
+        debugPrint('SendPhone complete');
+        // throw Exception('Something wrong');
+      });
+      emit(const AuthState.needOtp());
+    } on Exception catch (e, s) {
+      emit(AuthState.error(e, s));
+    }
+  }
 
-    on<AuthEventSendOtp>((event, emit) async {
-      emit(const AuthState.idle());
+  FutureOr<void> _onSendOtp(
+    AuthEventSendOtp _,
+    Emitter<AuthState> emit,
+  ) async {
+    emit(const AuthState.idle());
+    try {
       // TODO(any): handle incoming `AuthEventSendOtp` event
-      emit(const AuthState.success());
-    });
+      await Future.delayed(const Duration(seconds: 1), () {
+        debugPrint('SendOtp complete');
+        // throw Exception('Something wrong');
+      });
+      emit(const AuthState.successViaOtp());
+    } on Exception catch (e, s) {
+      emit(AuthState.error(e, s));
+    }
+  }
 
-    on<AuthEventLoginViaFacebook>((event, emit) async {
-      emit(const AuthState.idle());
-      // TODO(any): handle incoming `AuthEventLoginViaFacebook` event
-      emit(const AuthState.success());
-    });
+  FutureOr<void> _onLoginViaFacebook(
+    AuthEventLoginViaFacebook _,
+    Emitter<AuthState> emit,
+  ) async {
+    emit(const AuthState.idle());
+    // TODO(any): handle incoming `AuthEventLoginViaFacebook` event
+    emit(const AuthState.successViaSocial());
+  }
 
-    on<AuthEventLoginViaVkontakte>((event, emit) async {
-      emit(const AuthState.idle());
-      // TODO(any): handle incoming `AuthEventLoginViaVkontakte` event
-      emit(const AuthState.success());
-    });
-
-    // Пример обработчика в блоке
-    on<AuthEventLoginViaGoogle>((event, emit) async {
-      try {
-        emit(const AuthState.idle());
-
-        // TODO(Denis): отправить запрос в сервис google и передать токен в репозиторий
-
-        emit(const AuthState.success());
-      } on Exception catch (e, s) {
-        emit(AuthState.error(e, s));
+  FutureOr<void> _onLoginViaVkontakte(
+    AuthEventLoginViaVkontakte _,
+    Emitter<AuthState> emit,
+  ) async {
+    try {
+      final token = await vkAuth.logIn();
+      if (token.isNotEmpty) {
+        await authRepository.loginWithVk(
+          VkLoginRequest(token: token),
+        );
+        emit(const AuthState.successViaSocial());
       }
-    });
+    } on PlatformException catch (e, s) {
+      emit(AuthState.error(AuthorizationException(), s));
+    } on Exception catch (e, s) {
+      emit(AuthState.error(e, s));
+    }
+  }
 
-    on<AuthEventRegister>((event, emit) async {
+  FutureOr<void> _onLoginViaGoogle(
+    AuthEventLoginViaGoogle _,
+    Emitter<AuthState> emit,
+  ) async {
+    try {
       emit(const AuthState.idle());
-      debugPrint(
-        '${event.name}, ${event.surname}, ${event.email}, '
-        '${event.birthdate}, ${event.agree}',
+      final token = await googleAuth.signIn();
+      if (token == null) {
+        return;
+      }
+      await authRepository.loginWithGoogle(
+        GoogleLoginRequest(token: token),
       );
-      emit(const AuthState.success());
-    });
+      emit(const AuthState.successViaSocial());
+    } on PlatformException catch (e, s) {
+      emit(AuthState.error(AuthorizationException(), s));
+    } on Exception catch (e, s) {
+      emit(AuthState.error(e, s));
+    }
+  }
 
-    on<AuthEventLogout>((event, emit) async {
-      emit(const AuthState.idle());
-      // TODO(any): handle incoming `AuthEventLogout` event
-      emit(const AuthState.success());
-    });
+  FutureOr<void> _onRegister(
+    AuthEventRegister _,
+    Emitter<AuthState> emit,
+  ) async {
+    emit(const AuthState.idle());
+    // TODO(any): handle incoming `AuthEventRegister` event
+    emit(const AuthState.register());
+  }
+
+  FutureOr<void> _onAuthEventLogout(
+    AuthEventLogout _,
+    Emitter<AuthState> emit,
+  ) async {
+    emit(const AuthState.idle());
+    // TODO(any): handle incoming `AuthEventLogout` event
+    emit(const AuthState.logout());
   }
 }