+++ lib/src/navigation/app_router.dart
@@ -1,6 +1,7 @@
 import 'package:routemaster/routemaster.dart';
+import 'package:where_to_go_today/src/features/auth/code/code_route.dart';
 import 'package:where_to_go_today/src/features/auth/register/register_route.dart';
-import 'package:where_to_go_today/src/features/main/main_screen_route.dart';
+import 'package:where_to_go_today/src/features/auth/sign_in/sign_in_route.dart';
 import 'package:where_to_go_today/src/features/settings/ui/settings_route.dart';
 
 /// Класс, в котором описываем навигацию в приложении.
@@ -12,10 +13,12 @@ class AppRouter {
 
   static final routes = RouteMap(
     routes: {
-      // Временно заменил initalRoute для тестирования SignInScreen
-      initialRoute: (_) => MainScreenRoute(),
-      SettingsRoute.routeName: (_) => SettingsRoute(),
+      initialRoute: (_) => SignInRoute(),
+      SignInRoute.routeName: (_) => SignInRoute(),
+      '${CodeRoute.routeName}/:phone': (route) =>
+          CodeRoute(route.pathParameters['phone'] ?? ''),
       RegisterRoute.routeName: (_) => RegisterRoute(),
+      SettingsRoute.routeName: (_) => SettingsRoute(),
     },
   );
 }